CSP-2381 CSP-2382 Refactor V1 provider course model and controller usage
