+++ src/SFA.DAS.Roatp.Api.UnitTests/Controllers/ExternalReadControllers/ProvidersControllerTests/ProvidersControllerV1Tests.cs
@@ -10,7 +10,7 @@ using Microsoft.AspNetCore.Mvc;
 using Moq;
 using NUnit.Framework;
 using SFA.DAS.Roatp.Api.Controllers.ExternalReadControllers;
-using SFA.DAS.Roatp.Api.Models;
+using SFA.DAS.Roatp.Api.Models.V1;
 using SFA.DAS.Roatp.Application.Mediatr.Responses;
 using SFA.DAS.Roatp.Application.ProviderCourse.Queries.ExternalRead.GetProviderCourse;
 using SFA.DAS.Roatp.Application.ProviderCourse.Queries.GetAllProviderCourses;
@@ -21,7 +21,7 @@ using SFA.DAS.Roatp.Application.Providers.Queries.GetRegisteredProvider;
 using SFA.DAS.Roatp.Domain.Models;
 using SFA.DAS.Testing.AutoFixture;
 
-namespace SFA.DAS.Roatp.Api.UnitTests.Controllers.ExternalReadControllers;
+namespace SFA.DAS.Roatp.Api.UnitTests.Controllers.ExternalReadControllers.ProvidersControllerTests;
 
 public class ProvidersControllerV1Tests
 {
@@ -263,7 +263,7 @@ public class ProvidersControllerV1Tests
         var ok = actionResult as OkObjectResult;
         ok.Should().NotBeNull();
 
-        var returned = ok.Value as IList<ProviderCourseModelExternalModelV1>;
+        var returned = ok.Value as IList<ProviderCourseModelExternalModel>;
         returned.Should().NotBeNull();
         returned.Should().HaveCount(1);
     }
@@ -287,7 +287,7 @@ public class ProvidersControllerV1Tests
         mediatorMock.Setup(m => m.Send(It.Is<GetProviderCourseQuery>(q => q.Ukprn == ukprn && q.LarsCode == larsCode.ToString()), It.IsAny<CancellationToken>())).ReturnsAsync(new ValidatedResponse<ProviderCourseModel>(handlerResult));
         var result = await sut.GetProviderCourse(ukprn, larsCode);
 
-        var mappedResult = (ProviderCourseModelExternalModelV1)(ProviderCourseModelExternal)handlerResult;
+        var mappedResult = (ProviderCourseModelExternalModel)(ProviderCourseModelExternal)handlerResult;
         (result as OkObjectResult).Value.Should().BeEquivalentTo(mappedResult);
 
         mediatorMock.Verify(a =>
